declare group_pattern_util.h
