+++ paddle/cinn/adt/tree_test.cc
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 #include "paddle/cinn/adt/tree.h"
+#include "paddle/cinn/adt/tree_util.h"
 
 #include "gtest/gtest.h"
